Move TypeConverter to mlir/lib/Dialect/Tosa/Transforms/TosaTypeConverters.cpp
