+++ mlir/lib/Conversion/TosaToTensor/CMakeLists.txt
@@ -15,7 +15,6 @@ add_mlir_conversion_library(MLIRTosaToTensor
   MLIRIR
   MLIRPass
   MLIRTosaDialect
-  MLIRTosaToLinalg
   MLIRTosaTransforms
   MLIRSupport
   )
