+++ mlir/lib/Conversion/TosaToTensor/TosaToTensorPass.cpp
@@ -20,7 +20,6 @@
 #include "mlir/Pass/PassManager.h"
 #include "mlir/Transforms/DialectConversion.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
-#include <mlir/Conversion/TosaToLinalg/TosaToLinalg.h>
 
 namespace mlir {
 #define GEN_PASS_DEF_TOSATOTENSOR
@@ -44,7 +43,7 @@ public:
     target.addLegalDialect<tensor::TensorDialect>();
 
     TypeConverter converter;
-    mlir::tosa::populateTosaToLinalgTypeConversion(converter);
+    mlir::tosa::populateTosaTypeConversion(converter);
 
     mlir::tosa::populateTosaToTensorConversionPatterns(converter, &patterns);
